@@ -1,68 +1,24 @@
 package ar.security.controller;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
+@RestController
 public class HelloWorldController {
 	
-	@RequestMapping(value={"/", "home"}, method = RequestMethod.GET)
-	public String homePage(ModelMap model){
-		model.addAttribute("user", getPrincipal());
-		return "welcome";
+	@GetMapping("/home")
+	public String homePage(){
+		return "Hello";
 	}
 	
-	@RequestMapping(value="/admin", method=RequestMethod.GET)
-	public String adminPage(ModelMap model){
-		model.addAttribute("user", getPrincipal());
-		return "admin";
+	@GetMapping("/admin")
+	public String adminPage(){
+		return "Hello Admin";
 	}
 	
-	@RequestMapping(value="/dba", method= RequestMethod.GET)
-	public String dbaPage(ModelMap model){
-		model.addAttribute("user", getPrincipal());
-		return "dba";
+	@GetMapping(value="/dba")
+	public String dbaPage(){
+		return "Hello DBAdmin";
 	}
-	
-	@RequestMapping(value="/accessDenied", method= RequestMethod.GET)
-	public String accessDeniedPage(ModelMap model){
-		model.addAttribute("user", getPrincipal());
-		return "accessDenied";
-	}
-	
-	@RequestMapping(value="/logout", method= RequestMethod.GET)
-	public String logoutPage(HttpServletRequest request, HttpServletResponse response){
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		if(auth != null){
-			new SecurityContextLogoutHandler().logout(request, response, auth);
-		}
-		return  "redirect:/login?logout";
-	}
-	
-	@RequestMapping(value="/login", method= RequestMethod.GET)
-	public String loginPage(ModelMap model){
-		return "login";
-	}
-	
-	private String getPrincipal(){
-		String userName = null;
-		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		
-		if(principal instanceof UserDetails){
-			userName = ((UserDetails)principal).getUsername();
-		}else{
-			userName = principal.toString();
-		}
-		return userName;
-	}
+
 }
